examples/device-install: move the webhook URL field into the build section, where it is used
